+++ app/src/main/java/org/sopt/dosopttemplate/data/model/requestModel/RequestLoginDto.kt
@@ -1,14 +1,12 @@
-package org.sopt.dosopttemplate.model.requestModel
+package org.sopt.dosopttemplate.data.model.requestModel
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignupDto(
+data class RequestLoginDto(
     @SerialName("username")
     val username: String,
     @SerialName("password")
     val password: String,
-    @SerialName("nickname")
-    val nickname: String,
 )
