+++ app/src/main/java/org/sopt/dosopttemplate/data/model/requestModel/RequestSignupDto.kt
@@ -1,14 +1,14 @@
-package org.sopt.dosopttemplate.model.responseModel
+package org.sopt.dosopttemplate.data.model.requestModel
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseLoginDto(
-    @SerialName("id")
-    val id: Int,
+data class RequestSignupDto(
     @SerialName("username")
     val username: String,
+    @SerialName("password")
+    val password: String,
     @SerialName("nickname")
     val nickname: String,
 )
