+++ app/src/main/java/org/sopt/dosopttemplate/data/repository/UserRepository.kt
@@ -1,0 +1,12 @@
+package org.sopt.dosopttemplate.data.repository
+
+import org.sopt.dosopttemplate.data.service.UserService
+
+class UserRepository(
+    private val userService: UserService,
+) {
+    suspend fun loadUser(page: Int) =
+        runCatching {
+            userService.getUserList(page).data
+        }
+}
