+++ app/src/main/java/org/sopt/dosopttemplate/data/service/AuthService.kt
@@ -1,8 +1,8 @@
-package org.sopt.dosopttemplate.server.service
+package org.sopt.dosopttemplate.data.service
 
-import org.sopt.dosopttemplate.model.requestModel.RequestLoginDto
-import org.sopt.dosopttemplate.model.requestModel.RequestSignupDto
-import org.sopt.dosopttemplate.model.responseModel.ResponseLoginDto
+import org.sopt.dosopttemplate.data.model.requestModel.RequestLoginDto
+import org.sopt.dosopttemplate.data.model.requestModel.RequestSignupDto
+import org.sopt.dosopttemplate.data.model.responseModel.ResponseLoginDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
