+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/doandroid/DoAndroidFragment.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import org.sopt.dosopttemplate.base.BaseFragment
 import org.sopt.dosopttemplate.databinding.FragmentDoAndroidBinding
+import org.sopt.dosopttemplate.util.binding.BaseFragment
 
 class DoAndroidFragment : BaseFragment<FragmentDoAndroidBinding>() {
     override fun getFragmentBinding(
