+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/mypage/MyPageFragment.kt
@@ -8,8 +8,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import org.sopt.dosopttemplate.R
-import org.sopt.dosopttemplate.base.BaseFragment
 import org.sopt.dosopttemplate.databinding.FragmentMyPageBinding
+import org.sopt.dosopttemplate.util.binding.BaseFragment
 import org.sopt.dosopttemplate.utilprivate.makeToast
 
 class MyPageFragment : BaseFragment<FragmentMyPageBinding>() {
