+++ app/src/main/java/org/sopt/dosopttemplate/presentation/login/LoginActivity.kt
@@ -10,9 +10,9 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import org.sopt.dosopttemplate.R
+import org.sopt.dosopttemplate.data.model.User
+import org.sopt.dosopttemplate.data.model.responseModel.ResponseLoginDto
 import org.sopt.dosopttemplate.databinding.ActivityLoginBinding
-import org.sopt.dosopttemplate.model.User
-import org.sopt.dosopttemplate.model.responseModel.ResponseLoginDto
 import org.sopt.dosopttemplate.presentation.home.HomeActivity
 import org.sopt.dosopttemplate.presentation.signup.SignUpActivity
 import org.sopt.dosopttemplate.utilprivate.makeToast
