--- app/src/main/java/org/sopt/dosopttemplate/repository/UserRepository.kt
@@ -1,12 +1,0 @@
-package org.sopt.dosopttemplate.repository
-
-import org.sopt.dosopttemplate.server.service.UserService
-
-class UserRepository(
-    private val userService: UserService
-) {
-    suspend fun loadUser(page: Int) =
-        runCatching {
-            userService.getUserList(page).data
-        }
-}
