+++ app/src/main/java/org/sopt/dosopttemplate/util/binding/BaseActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.base
+package org.sopt.dosopttemplate.util.binding
 
 import android.app.Activity
 import android.os.Bundle
@@ -8,7 +8,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.viewbinding.ViewBinding
 
 abstract class BaseActivity<T : ViewBinding>(
-    val bindingFactory: (LayoutInflater) -> T
+    val bindingFactory: (LayoutInflater) -> T,
 ) : AppCompatActivity() {
     private var _binding: T? = null
     val binding get() = _binding!!
